Fix Wshorten-64-to-32 warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38352 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/samplerdrv/file_drv.c
@@ -1715,7 +1715,7 @@ static int handle_mp3_file(int channels)
     buffer_size = mpg123_length(mh);
 
     lib_free(file_buffer);
-    file_size = buffer_size * 2 * mp3_channels;
+    file_size = (unsigned int)(buffer_size * 2 * mp3_channels);
     file_buffer = lib_malloc(file_size);
 
     mp3_err = mpg123_read(mh, file_buffer, file_size, &done);
@@ -1728,7 +1728,7 @@ static int handle_mp3_file(int channels)
 
     sound_audio_type = AUDIO_TYPE_PCM;
     sound_audio_channels = mp3_channels;
-    sound_audio_rate = mp3_rate;
+    sound_audio_rate = (unsigned int)mp3_rate;
     sound_audio_bits = 16;
 
     return convert_pcm_buffer(file_size, channels);
